Refactor smart budgeting module into modular components
