+++ src/views/SmartBudgetingView.tsx
@@ -1,2335 +1,36 @@
-import { FormEvent, useEffect, useMemo, useState } from 'react';
-import { addMonths, format, formatISO, parseISO } from 'date-fns';
-import { useFinancialStore } from '../store/FinancialStoreProvider';
-import type { Category, PlannedExpenseItem, Transaction } from '../types';
-
-function formatCurrency(value: number) {
-  return new Intl.NumberFormat('en-IN', { style: 'currency', currency: 'INR', maximumFractionDigits: 0 }).format(value);
-}
-
-function monthKey(date?: string | null) {
-  return date ? date.slice(0, 7) : '';
-}
-
-function yearKey(date?: string | null) {
-  return date ? date.slice(0, 4) : '';
-}
-
-function formatMonthLabel(month: string) {
-  try {
-    return format(parseISO(`${month}-01`), 'MMMM yyyy');
-  } catch {
-    return month;
-  }
-}
-
-function defaultDueDateForPeriod(viewMode: 'monthly' | 'yearly', month: string, year: string) {
-  const today = new Date();
-  if (viewMode === 'monthly') {
-    const [yearPart, monthPart] = month.split('-');
-    const safeYear = Number.parseInt(yearPart, 10);
-    const safeMonth = Number.parseInt(monthPart, 10) - 1;
-    if (Number.isNaN(safeYear) || Number.isNaN(safeMonth)) {
-      return formatISO(today, { representation: 'date' });
-    }
-    const lastDayOfMonth = new Date(safeYear, safeMonth + 1, 0).getDate();
-    const preferredDay = Math.min(today.getDate(), lastDayOfMonth);
-    const defaultDate = new Date(safeYear, safeMonth, preferredDay);
-    return formatISO(defaultDate, { representation: 'date' });
-  }
-
-  const safeYear = Number.parseInt(year, 10);
-  if (Number.isNaN(safeYear)) {
-    return formatISO(today, { representation: 'date' });
-  }
-  const defaultDate = new Date(safeYear, today.getMonth(), today.getDate());
-  return formatISO(defaultDate, { representation: 'date' });
-}
-
-const PRIORITY_TOKEN_STYLES: Record<PlannedExpenseItem['priority'], { label: string; badgeClass: string }> = {
-  high: { label: 'High priority', badgeClass: 'bg-danger/20 text-danger' },
-  medium: { label: 'Medium priority', badgeClass: 'bg-warning/20 text-warning' },
-  low: { label: 'Low priority', badgeClass: 'bg-slate-800 text-slate-300' }
-};
-
-const PRIORITY_OPTIONS: Array<{ value: PlannedExpenseItem['priority']; label: string }> = [
-  { value: 'high', label: 'High priority' },
-  { value: 'medium', label: 'Medium priority' },
-  { value: 'low', label: 'Low priority' }
-];
-
-const PRIORITY_ORDER: PlannedExpenseItem['priority'][] = ['high', 'medium', 'low'];
+import { CategoryInspector } from '../features/smart-budgeting/organisms/CategoryInspector';
+import { CategoryNavigator } from '../features/smart-budgeting/organisms/CategoryNavigator';
+import { NavigatorFilters } from '../features/smart-budgeting/organisms/NavigatorFilters';
+import { PlannedExpensesDialog } from '../features/smart-budgeting/organisms/PlannedExpensesDialog';
+import { SummaryGrid } from '../features/smart-budgeting/organisms/SummaryGrid';
+import { SummaryHeaderControls } from '../features/smart-budgeting/organisms/SummaryHeaderControls';
+import { useSmartBudgetingController } from '../features/smart-budgeting/hooks/useSmartBudgetingController';
 
 export function SmartBudgetingView() {
-  const {
-    plannedExpenses,
-    categories,
-    transactions,
-    addPlannedExpense,
-    updatePlannedExpense,
-    updateCategory,
-    deletePlannedExpense,
-    addCategory
-  } = useFinancialStore();
-  const now = new Date();
-  const defaultMonth = format(now, 'yyyy-MM');
-  const defaultYear = format(now, 'yyyy');
+  const controller = useSmartBudgetingController();
+  const { dialog, utils, period, categories, overview, inspector, editing } = controller;
 
-  const expenseCategories = useMemo(
-    () => categories.filter((category) => category.type === 'expense'),
-    [categories]
-  );
-
-  const [viewMode, setViewMode] = useState<'monthly' | 'yearly'>('monthly');
-  const [selectedMonth, setSelectedMonth] = useState(defaultMonth);
-  const [selectedYear, setSelectedYear] = useState(defaultYear);
-  const [selectedCategoryId, setSelectedCategoryId] = useState<'all' | string>('all');
-  const periodLabel = useMemo(
-    () => (viewMode === 'monthly' ? formatMonthLabel(selectedMonth) : selectedYear),
-    [viewMode, selectedMonth, selectedYear]
-  );
-
-  const handleViewModeChange = (mode: 'monthly' | 'yearly') => {
-    if (mode === viewMode) {
-      return;
-    }
-    if (mode === 'yearly') {
-      const derivedYear = selectedMonth.slice(0, 4);
-      if (derivedYear) {
-        setSelectedYear(derivedYear);
-      }
-    }
-    setViewMode(mode);
-  };
-
-  const shiftPeriod = (direction: -1 | 1) => {
-    if (viewMode === 'monthly') {
-      try {
-        const baseDate = parseISO(`${selectedMonth}-01`);
-        if (Number.isNaN(baseDate.getTime())) {
-          throw new Error('Invalid month');
-        }
-        const nextDate = addMonths(baseDate, direction);
-        setSelectedMonth(format(nextDate, 'yyyy-MM'));
-      } catch {
-        const fallbackDate = addMonths(new Date(), direction);
-        setSelectedMonth(format(fallbackDate, 'yyyy-MM'));
-      }
-      return;
-    }
-
-    const safeYear = Number.parseInt(selectedYear, 10);
-    if (Number.isNaN(safeYear)) {
-      const fallbackYear = Number.parseInt(defaultYear, 10) + direction;
-      setSelectedYear(String(fallbackYear));
-      return;
-    }
-    setSelectedYear(String(safeYear + direction));
-  };
-
-  const goToPreviousPeriod = () => shiftPeriod(-1);
-  const goToNextPeriod = () => shiftPeriod(1);
-
-  type PlannedExpenseDraft = {
-    id: string;
-    name: string;
-    amount: string;
-    dueDate: string;
-    hasDueDate: boolean;
-    categoryId: string;
-    priority: PlannedExpenseItem['priority'];
-  };
-
-  const generateEntryId = () => Math.random().toString(36).slice(2);
-
-  const resolveDefaultDueDate = () => {
-    try {
-      if (viewMode === 'monthly') {
-        return formatISO(parseISO(`${selectedMonth}-01`), { representation: 'date' });
-      }
-      if (viewMode === 'yearly') {
-        return formatISO(parseISO(`${selectedYear}-01-01`), { representation: 'date' });
-      }
-    } catch {
-      // Fall through to the generic fallback below when parsing fails.
-    }
-    return formatISO(addMonths(new Date(), 1), { representation: 'date' });
-  };
-
-  const createEmptyEntry = (categoryId?: string): PlannedExpenseDraft => ({
-    id: generateEntryId(),
-    name: '',
-    amount: '',
-    dueDate: resolveDefaultDueDate(),
-    hasDueDate: true,
-    categoryId: categoryId ?? expenseCategories[0]?.id ?? '',
-    priority: 'medium'
-  });
-
-  const [plannedEntries, setPlannedEntries] = useState<PlannedExpenseDraft[]>(() => [createEmptyEntry()]);
-  const [isAddExpenseDialogOpen, setIsAddExpenseDialogOpen] = useState(false);
-  const [isSubmittingPlannedExpenses, setIsSubmittingPlannedExpenses] = useState(false);
-  const [categoryCreationTargetId, setCategoryCreationTargetId] = useState<string | null>(null);
-  const [newCategoryName, setNewCategoryName] = useState('');
-  const [hasAttemptedSubmit, setHasAttemptedSubmit] = useState(false);
-
-  useEffect(() => {
-    if (viewMode === 'monthly') {
-      const derivedYear = selectedMonth.slice(0, 4);
-      if (derivedYear && derivedYear !== selectedYear) {
-        setSelectedYear(derivedYear);
-      }
-    }
-  }, [selectedMonth, selectedYear, viewMode]);
-
-  useEffect(() => {
-    if (viewMode === 'yearly') {
-      setSelectedMonth((previous) => {
-        const parts = previous.split('-');
-        const monthPart = parts[1] ? parts[1].padStart(2, '0') : format(new Date(), 'MM');
-        const nextValue = `${selectedYear}-${monthPart}`;
-        return nextValue === previous ? previous : nextValue;
-      });
-    }
-  }, [selectedYear, viewMode]);
-
-  useEffect(() => {
-    if (expenseCategories.length === 0) {
-      setPlannedEntries((previous) =>
-        previous.map((entry) => ({ ...entry, categoryId: '' }))
-      );
-      return;
-    }
-    setPlannedEntries((previous) =>
-      previous.map((entry) =>
-        expenseCategories.some((category) => category.id === entry.categoryId)
-          ? entry
-          : { ...entry, categoryId: expenseCategories[0].id }
-      )
-    );
-  }, [expenseCategories]);
-
-  const categoryLookup = useMemo(() => new Map(categories.map((category) => [category.id, category])), [categories]);
-  const categoryParentMap = useMemo(() => {
-    const map = new Map<string, string | null>();
-    expenseCategories.forEach((category) => {
-      if (!category.parentId) {
-        map.set(category.id, null);
-        return;
-      }
-      const parent = categoryLookup.get(category.parentId);
-      if (parent && parent.type === 'expense') {
-        map.set(category.id, parent.id);
-      } else {
-        map.set(category.id, null);
-      }
-    });
-    return map;
-  }, [expenseCategories, categoryLookup]);
-  useEffect(() => {
-    if (selectedCategoryId !== 'all' && !categoryLookup.has(selectedCategoryId)) {
-      setSelectedCategoryId('all');
-    }
-  }, [categoryLookup, selectedCategoryId]);
-
-  const expenseDescendantsMap = useMemo(() => {
-    const childrenMap = new Map<string, string[]>();
-    expenseCategories.forEach((category) => {
-      if (!category.parentId) return;
-      const parent = categoryLookup.get(category.parentId);
-      if (!parent || parent.type !== 'expense') return;
-      const list = childrenMap.get(category.parentId) ?? [];
-      list.push(category.id);
-      childrenMap.set(category.parentId, list);
-    });
-    const map = new Map<string, Set<string>>();
-    const visit = (id: string): Set<string> => {
-      const existing = map.get(id);
-      if (existing) return existing;
-      const set = new Set<string>([id]);
-      (childrenMap.get(id) ?? []).forEach((childId) => {
-        visit(childId).forEach((descId) => set.add(descId));
-      });
-      map.set(id, set);
-      return set;
-    };
-    expenseCategories.forEach((category) => visit(category.id));
-    return map;
-  }, [expenseCategories, categoryLookup]);
-
-  const allExpenseIdsSet = useMemo(() => new Set(expenseCategories.map((category) => category.id)), [expenseCategories]);
-
-  const periodPlannedExpenses = useMemo(
-    () =>
-      plannedExpenses.filter((item) => {
-        if (item.status === 'cancelled') {
-          return false;
-        }
-        const referenceDate = item.dueDate ?? item.createdAt;
-        return viewMode === 'monthly'
-          ? monthKey(referenceDate) === selectedMonth
-          : yearKey(referenceDate) === selectedYear;
-      }),
-    [plannedExpenses, viewMode, selectedMonth, selectedYear]
-  );
-
-  const periodTransactions = useMemo(
-    () =>
-      transactions.filter(
-        (txn) =>
-          txn.amount < 0 &&
-          (viewMode === 'monthly' ? monthKey(txn.date) === selectedMonth : yearKey(txn.date) === selectedYear)
-      ),
-    [transactions, viewMode, selectedMonth, selectedYear]
-  );
-
-  const resolveCategoryIds = (categoryId: 'all' | string) => {
-    if (categoryId === 'all') return allExpenseIdsSet;
-    return expenseDescendantsMap.get(categoryId) ?? new Set<string>([categoryId]);
-  };
-
-  const computeTotals = (categoryId: 'all' | string) => {
-    const ids = resolveCategoryIds(categoryId);
-    const plannedItems = periodPlannedExpenses.filter((item) => ids.has(item.categoryId));
-    const plannedFromItems = plannedItems.reduce((sum, item) => sum + item.plannedAmount, 0);
-    const actualEntries = periodTransactions.filter(
-      (txn) => txn.categoryId && ids.has(txn.categoryId)
-    );
-    const actualTotal = actualEntries.reduce((sum, txn) => sum + Math.abs(txn.amount), 0);
-    const budgetTotal = Array.from(ids).reduce((sum, id) => {
-      const category = categoryLookup.get(id);
-      if (!category) return sum;
-      const budget = viewMode === 'monthly' ? category.budgets?.monthly : category.budgets?.yearly;
-      return sum + (budget ?? 0);
-    }, 0);
-    return {
-      plannedItems,
-      actualEntries,
-      plannedFromItems,
-      budgetTotal,
-      totalPlanned: plannedFromItems + budgetTotal,
-      actualTotal
-    };
-  };
-
-  const totalsForAll = useMemo(
-    () => computeTotals('all'),
-    [periodPlannedExpenses, periodTransactions, viewMode, categoryLookup, expenseDescendantsMap, allExpenseIdsSet]
-  );
-
-  const totalsForSelected = useMemo(
-    () => computeTotals(selectedCategoryId),
-    [selectedCategoryId, periodPlannedExpenses, periodTransactions, viewMode, categoryLookup, expenseDescendantsMap, allExpenseIdsSet]
-  );
-
-  type PlannedExpenseSpendingHealth = 'not-spent' | 'under' | 'over';
-
-  type PlannedExpenseDetail = {
-    item: PlannedExpenseItem;
-    match?: Transaction;
-    actual: number;
-    variance: number;
-    status: PlannedExpenseSpendingHealth;
-    remainder: number;
-    priority: PlannedExpenseItem['priority'];
-  };
-
-  type CategoryNode = Category & { children: CategoryNode[] };
-
-  const [expandedCategories, setExpandedCategories] = useState<Record<string, boolean>>({});
-  const [editingItemId, setEditingItemId] = useState<string | null>(null);
-  const [editDraft, setEditDraft] = useState<{
-    categoryId: string;
-    plannedAmount: string;
-    actualAmount: string;
-    remainderAmount: string;
-    dueDate: string;
-    hasDueDate: boolean;
-    priority: PlannedExpenseItem['priority'];
-  }>({
-    categoryId: '',
-    plannedAmount: '',
-    actualAmount: '',
-    remainderAmount: '',
-    dueDate: '',
-    hasDueDate: true,
-    priority: 'medium'
-  });
-  const [savingItemId, setSavingItemId] = useState<string | null>(null);
-  const [quickActualDrafts, setQuickActualDrafts] = useState<Record<string, string>>({});
-  const [quickActualSavingId, setQuickActualSavingId] = useState<string | null>(null);
-  const [navigatorFilter, setNavigatorFilter] = useState<'all' | PlannedExpenseSpendingHealth>('all');
-  const [navigatorView, setNavigatorView] = useState<'category' | 'priority'>('category');
-  const [categorySearchTerm, setCategorySearchTerm] = useState('');
-  const [focusedCategoryId, setFocusedCategoryId] = useState<string | null>(null);
-  const [budgetDraft, setBudgetDraft] = useState('');
-  const [isSavingBudget, setIsSavingBudget] = useState(false);
-  const [budgetError, setBudgetError] = useState<string | null>(null);
-  const [budgetFeedback, setBudgetFeedback] = useState<'saved' | 'cleared' | null>(null);
-  const navigatorFilterOptions: Array<{ key: 'all' | PlannedExpenseSpendingHealth; label: string }> = [
-    { key: 'all', label: 'All statuses' },
-    { key: 'over', label: 'Overspending' },
-    { key: 'under', label: 'Under budget' },
-    { key: 'not-spent', label: 'Awaiting spend' }
-  ];
-  const navigatorViewOptions: Array<{ key: 'category' | 'priority'; label: string }> = [
-    { key: 'category', label: 'Category view' },
-    { key: 'priority', label: 'Priority view' }
-  ];
-  const normalisedSearchTerm = categorySearchTerm.trim().toLowerCase();
-
-  useEffect(() => {
-    setExpandedCategories((previous) => {
-      const next: Record<string, boolean> = {};
-      expenseCategories.forEach((category) => {
-        const parent = expenseCategories.find((candidate) => candidate.id === category.parentId);
-        const isRoot = !category.parentId || !parent;
-        next[category.id] = previous[category.id] ?? isRoot;
-      });
-      return next;
-    });
-  }, [expenseCategories]);
-
-  useEffect(() => {
-    setBudgetError(null);
-    setBudgetFeedback(null);
-  }, [focusedCategoryId, viewMode]);
-
-  const plannedExpenseDetails = useMemo<PlannedExpenseDetail[]>(() => {
-    return periodPlannedExpenses
-      .map((item) => {
-        const match = periodTransactions.find(
-          (txn) =>
-            txn.categoryId === item.categoryId &&
-            txn.amount < 0 &&
-            Math.abs(Math.abs(txn.amount) - item.plannedAmount) <= Math.max(500, item.plannedAmount * 0.1)
-        );
-        const matchedAmount = match ? Math.abs(match.amount) : 0;
-        const manualActual =
-          typeof item.actualAmount === 'number' && !Number.isNaN(item.actualAmount) ? item.actualAmount : undefined;
-        const remainderOverride =
-          typeof item.remainderAmount === 'number' && !Number.isNaN(item.remainderAmount)
-            ? Math.max(item.remainderAmount, 0)
-            : null;
-        const actualAmount =
-          manualActual ??
-          (remainderOverride !== null ? Math.max(item.plannedAmount - remainderOverride, 0) : matchedAmount);
-        const variance = item.plannedAmount - actualAmount;
-        const remainder = remainderOverride !== null ? remainderOverride : variance;
-        const status: PlannedExpenseSpendingHealth =
-          actualAmount === 0 ? 'not-spent' : variance >= 0 ? 'under' : 'over';
-        return {
-          item: { ...item, priority: item.priority ?? 'medium' },
-          match,
-          actual: actualAmount,
-          variance,
-          status,
-          remainder,
-          priority: item.priority ?? 'medium'
-        } satisfies PlannedExpenseDetail;
-      })
-      .sort((a, b) => {
-        const hasDueA = Boolean(a.item.dueDate);
-        const hasDueB = Boolean(b.item.dueDate);
-        if (hasDueA && hasDueB) {
-          const diff = new Date(a.item.dueDate!).getTime() - new Date(b.item.dueDate!).getTime();
-          if (diff !== 0) {
-            return diff;
-          }
-        } else if (hasDueA !== hasDueB) {
-          return hasDueA ? -1 : 1;
-        }
-        const createdDiff = new Date(a.item.createdAt).getTime() - new Date(b.item.createdAt).getTime();
-        if (createdDiff !== 0) {
-          return createdDiff;
-        }
-        return a.item.name.localeCompare(b.item.name);
-      });
-  }, [periodPlannedExpenses, periodTransactions]);
-
-  const expenseCategoryTree = useMemo<CategoryNode[]>(() => {
-    const nodes = new Map<string, CategoryNode>();
-    expenseCategories.forEach((category) => {
-      nodes.set(category.id, { ...category, children: [] });
-    });
-    const roots: CategoryNode[] = [];
-    nodes.forEach((node) => {
-      if (node.parentId && nodes.has(node.parentId)) {
-        nodes.get(node.parentId)!.children.push(node);
-      } else {
-        roots.push(node);
-      }
-    });
-    const sortNodes = (list: CategoryNode[]) => {
-      list.sort((a, b) => a.name.localeCompare(b.name));
-      list.forEach((child) => sortNodes(child.children));
-    };
-    sortNodes(roots);
-    return roots;
-  }, [expenseCategories]);
-
-  const categoryOptions = useMemo(() => {
-    const options: Array<{ id: string; label: string }> = [];
-    const walk = (nodes: CategoryNode[], depth = 0) => {
-      nodes.forEach((node) => {
-        const prefix = depth === 0 ? '' : `${'—'.repeat(depth)} `;
-        options.push({ id: node.id, label: `${prefix}${node.name}` });
-        if (node.children.length > 0) {
-          walk(node.children, depth + 1);
-        }
-      });
-    };
-    walk(expenseCategoryTree);
-    return options;
-  }, [expenseCategoryTree]);
-
-  const expenseCategoryIds = useMemo(() => new Set(expenseCategories.map((category) => category.id)), [expenseCategories]);
-
-  const { itemsByCategory, categorySummaries } = useMemo(() => {
-    const byCategory = new Map<string, PlannedExpenseDetail[]>();
-    plannedExpenseDetails.forEach((detail) => {
-      const list = byCategory.get(detail.item.categoryId) ?? [];
-      list.push(detail);
-      byCategory.set(detail.item.categoryId, list);
-    });
-
-    const summaries = new Map<
-      string,
-      { planned: number; actual: number; variance: number; itemCount: number; remainder: number }
-    >();
-
-    expenseCategories.forEach((category) => {
-      const ids = expenseDescendantsMap.get(category.id) ?? new Set<string>([category.id]);
-      let planned = 0;
-      let actual = 0;
-      let remainder = 0;
-      let itemCount = 0;
-      ids.forEach((id) => {
-        const entries = byCategory.get(id);
-        if (!entries) return;
-        itemCount += entries.length;
-        entries.forEach((detail) => {
-          planned += detail.item.plannedAmount;
-          actual += detail.actual;
-          remainder += detail.remainder;
-        });
-      });
-      summaries.set(category.id, {
-        planned,
-        actual,
-        variance: planned - actual,
-        itemCount,
-        remainder
-      });
-    });
-
-    return { itemsByCategory: byCategory, categorySummaries: summaries };
-  }, [plannedExpenseDetails, expenseCategories, expenseDescendantsMap]);
-
-  const categoriesWithContent = useMemo(
-    () =>
-      expenseCategories.filter((category) => (categorySummaries.get(category.id)?.itemCount ?? 0) > 0),
-    [expenseCategories, categorySummaries]
-  );
-
-  useEffect(() => {
-    if (editingItemId && !periodPlannedExpenses.some((item) => item.id === editingItemId)) {
-      setEditingItemId(null);
-    }
-  }, [editingItemId, periodPlannedExpenses]);
-
-  useEffect(() => {
-    const fallbackCategories =
-      categoriesWithContent.length > 0 ? categoriesWithContent : expenseCategories;
-    if (fallbackCategories.length === 0) {
-      setFocusedCategoryId(null);
-      return;
-    }
-    if (!focusedCategoryId || !fallbackCategories.some((category) => category.id === focusedCategoryId)) {
-      setFocusedCategoryId(fallbackCategories[0].id);
-    }
-  }, [categoriesWithContent, expenseCategories, focusedCategoryId]);
-
-  useEffect(() => {
-    if (!focusedCategoryId) {
-      setBudgetDraft('');
-      return;
-    }
-    const category = categoryLookup.get(focusedCategoryId);
-    if (!category) {
-      setBudgetDraft('');
-      return;
-    }
-    const budgetValue = viewMode === 'monthly' ? category.budgets?.monthly : category.budgets?.yearly;
-    setBudgetDraft(
-      typeof budgetValue === 'number' && !Number.isNaN(budgetValue) ? String(budgetValue) : ''
-    );
-  }, [focusedCategoryId, categoryLookup, viewMode]);
-
-  const spendingBadgeStyles: Record<PlannedExpenseSpendingHealth, { label: string; badgeClass: string; toneClass: string }> = {
-    'not-spent': {
-      label: 'Awaiting spend',
-      badgeClass: 'bg-slate-800 text-slate-300',
-      toneClass: 'text-slate-300'
-    },
-    under: {
-      label: 'Spent wisely',
-      badgeClass: 'bg-success/20 text-success',
-      toneClass: 'text-success'
-    },
-    over: {
-      label: 'Overspent',
-      badgeClass: 'bg-danger/20 text-danger',
-      toneClass: 'text-danger'
-    }
-  };
-
-  const progressColorByStatus: Record<PlannedExpenseSpendingHealth, string> = {
-    'not-spent': '#38bdf8',
-    under: '#10b981',
-    over: '#ef4444'
-  };
-
-  const uncategorisedDetails = useMemo(
-    () => plannedExpenseDetails.filter((detail) => !expenseCategoryIds.has(detail.item.categoryId)),
-    [plannedExpenseDetails, expenseCategoryIds]
-  );
-
-  const visibleUncategorisedDetails = useMemo(
-    () =>
-      uncategorisedDetails.filter((detail) => {
-        const matchesFilter = navigatorFilter === 'all' || detail.status === navigatorFilter;
-        const matchesSearch =
-          normalisedSearchTerm === '' || detail.item.name.toLowerCase().includes(normalisedSearchTerm);
-        return matchesFilter && matchesSearch;
-      }),
-    [uncategorisedDetails, navigatorFilter, normalisedSearchTerm]
-  );
-
-  const filteredPriorityDetails = useMemo(() => {
-    return plannedExpenseDetails.filter((detail) => {
-      const matchesFilter = navigatorFilter === 'all' || detail.status === navigatorFilter;
-      if (!matchesFilter) {
-        return false;
-      }
-      if (normalisedSearchTerm === '') {
-        return true;
-      }
-      const categoryName = categoryLookup.get(detail.item.categoryId)?.name?.toLowerCase() ?? 'uncategorised';
-      return (
-        detail.item.name.toLowerCase().includes(normalisedSearchTerm) || categoryName.includes(normalisedSearchTerm)
-      );
-    });
-  }, [plannedExpenseDetails, navigatorFilter, normalisedSearchTerm, categoryLookup]);
-
-  const priorityGroups = useMemo(() => {
-    const groups: Record<PlannedExpenseItem['priority'], PlannedExpenseDetail[]> = {
-      high: [],
-      medium: [],
-      low: []
-    };
-    const timeForDetail = (detail: PlannedExpenseDetail) =>
-      detail.item.dueDate ? new Date(detail.item.dueDate).getTime() : Number.POSITIVE_INFINITY;
-    filteredPriorityDetails.forEach((detail) => {
-      const priority = detail.priority ?? 'medium';
-      groups[priority].push(detail);
-    });
-    PRIORITY_ORDER.forEach((priority) => {
-      groups[priority].sort((a, b) => {
-        const dueComparison = timeForDetail(a) - timeForDetail(b);
-        if (dueComparison !== 0) {
-          return dueComparison;
-        }
-        return a.item.name.localeCompare(b.item.name);
-      });
-    });
-    return groups;
-  }, [filteredPriorityDetails]);
-
-  const overallSummary = useMemo(() => {
-    const planned = plannedExpenseDetails.reduce((sum, detail) => sum + detail.item.plannedAmount, 0);
-    const actual = plannedExpenseDetails.reduce((sum, detail) => sum + detail.actual, 0);
-    const variance = planned - actual;
-    const status: PlannedExpenseSpendingHealth =
-      plannedExpenseDetails.length === 0 || actual === 0
-        ? 'not-spent'
-        : variance >= 0
-        ? 'under'
-        : 'over';
-    return { planned, actual, variance, status };
-  }, [plannedExpenseDetails]);
-  const overallUtilisationPercent = overallSummary.planned <= 0 ? 0 : Math.round((overallSummary.actual / overallSummary.planned) * 100);
-  const overallUtilisationWidth = Math.max(0, Math.min(100, overallUtilisationPercent));
-
-  const overspendingCategories = useMemo(() => {
-    const list: Array<{
-      category: Category;
-      summary: { planned: number; actual: number; variance: number; itemCount: number };
-    }> = [];
-    expenseCategories.forEach((category) => {
-      const summary = categorySummaries.get(category.id);
-      if (summary && summary.itemCount > 0 && summary.variance < 0) {
-        list.push({ category, summary });
-      }
-    });
-    return list.sort((a, b) => a.summary.variance - b.summary.variance).slice(0, 3);
-  }, [expenseCategories, categorySummaries]);
-  const focusedCategory = focusedCategoryId ? categoryLookup.get(focusedCategoryId) ?? null : null;
-  const inspectorCategory = focusedCategory ?? undefined;
-  const inspectorDetails = useMemo(() => {
-    if (!focusedCategoryId) {
-      return [] as PlannedExpenseDetail[];
-    }
-    const ids = expenseDescendantsMap.get(focusedCategoryId) ?? new Set<string>([focusedCategoryId]);
-    return plannedExpenseDetails.filter((detail) => ids.has(detail.item.categoryId));
-  }, [focusedCategoryId, plannedExpenseDetails, expenseDescendantsMap]);
-  const inspectorOverspendingItems = useMemo(() => {
-    const overs = inspectorDetails.filter((detail) => detail.status === 'over');
-    overs.sort((a, b) => a.variance - b.variance);
-    return overs.slice(0, 3);
-  }, [inspectorDetails]);
-  const inspectorUpcomingItems = useMemo(() => {
-    const upcoming = inspectorDetails.filter(
-      (detail) => detail.status !== 'over' && detail.item.dueDate
-    );
-    upcoming.sort((a, b) => new Date(a.item.dueDate!).getTime() - new Date(b.item.dueDate!).getTime());
-    return upcoming.slice(0, 3);
-  }, [inspectorDetails]);
-  const inspectorInProgressItems = useMemo(() => {
-    const inProgress = inspectorDetails.filter(
-      (detail) => detail.item.status === 'pending' || detail.item.status === 'purchased'
-    );
-    inProgress.sort((a, b) => new Date(a.item.dueDate).getTime() - new Date(b.item.dueDate).getTime());
-    return inProgress.slice(0, 3);
-  }, [inspectorDetails]);
-  const inspectorCompletedItems = useMemo(() => {
-    const completed = inspectorDetails.filter((detail) => detail.item.status === 'reconciled');
-    completed.sort((a, b) => new Date(b.item.dueDate).getTime() - new Date(a.item.dueDate).getTime());
-    return completed.slice(0, 3);
-  }, [inspectorDetails]);
-
-  const updateAllCategoryExpansion = (expanded: boolean) => {
-    const next: Record<string, boolean> = {};
-    const visit = (nodes: CategoryNode[]) => {
-      nodes.forEach((node) => {
-        next[node.id] = expanded;
-        if (node.children.length > 0) {
-          visit(node.children);
-        }
-      });
-    };
-    visit(expenseCategoryTree);
-    setExpandedCategories(next);
-  };
-
-  const expandAllCategories = () => {
-    updateAllCategoryExpansion(true);
-  };
-
-  const collapseAllCategories = () => {
-    updateAllCategoryExpansion(false);
-  };
-
-  const statusBadge = (status: PlannedExpenseItem['status']) => {
-    const baseClass = 'inline-flex items-center rounded-full px-2 py-0.5 text-[10px] font-semibold uppercase tracking-wide';
-    switch (status) {
-      case 'pending':
-        return <span className={`${baseClass} bg-warning/20 text-warning`}>Pending</span>;
-      case 'purchased':
-        return <span className={`${baseClass} bg-sky-500/20 text-sky-300`}>Purchased</span>;
-      case 'reconciled':
-        return <span className={`${baseClass} bg-success/20 text-success`}>Reconciled</span>;
-      case 'cancelled':
-      default:
-        return <span className={`${baseClass} bg-danger/20 text-danger`}>Cancelled</span>;
-    }
-  };
-
-  const resetPlannedEntries = () => setPlannedEntries([createEmptyEntry()]);
-
-  const isEntryValid = (entry: PlannedExpenseDraft) => {
-    if (!entry.name.trim()) return false;
-    if (entry.amount.trim() === '') return false;
-    const numericAmount = Number(entry.amount);
-    if (Number.isNaN(numericAmount) || numericAmount < 0) return false;
-    if (entry.hasDueDate && !entry.dueDate) return false;
-    if (!entry.categoryId) return false;
-    return true;
-  };
-
-  const handleEntryChange = (id: string, patch: Partial<PlannedExpenseDraft>) => {
-    setPlannedEntries((previous) => previous.map((entry) => (entry.id === id ? { ...entry, ...patch } : entry)));
-  };
-
-  const handleAddEntryRow = () => {
-    setPlannedEntries((previous) => [...previous, createEmptyEntry()]);
-  };
-
-  const handleRemoveEntryRow = (id: string) => {
-    setPlannedEntries((previous) => {
-      if (previous.length <= 1) {
-        return [createEmptyEntry()];
-      }
-      return previous.filter((entry) => entry.id !== id);
-    });
-    setCategoryCreationTargetId((previous) => (previous === id ? null : previous));
-  };
-
-  const handleToggleCategoryCreation = (id: string) => {
-    setCategoryCreationTargetId((previous) => {
-      const next = previous === id ? null : id;
-      setNewCategoryName('');
-      return next;
-    });
-  };
-
-  const handleSubmitPlannedExpenses = async (event: FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
-    setHasAttemptedSubmit(true);
-    if (expenseCategories.length === 0) {
-      return;
-    }
-    const hasInvalidEntries = plannedEntries.some((entry) => !isEntryValid(entry));
-    if (hasInvalidEntries) {
-      return;
-    }
-    setIsSubmittingPlannedExpenses(true);
-    try {
-      for (const entry of plannedEntries) {
-        await addPlannedExpense({
-          name: entry.name.trim(),
-          plannedAmount: Number(entry.amount),
-          categoryId: entry.categoryId,
-          dueDate: entry.hasDueDate ? entry.dueDate : null,
-          priority: entry.priority,
-          status: 'pending'
-        });
-      }
-      resetPlannedEntries();
-      setCategoryCreationTargetId(null);
-      setNewCategoryName('');
-      setHasAttemptedSubmit(false);
-      setIsAddExpenseDialogOpen(false);
-    } finally {
-      setIsSubmittingPlannedExpenses(false);
-    }
-  };
-
-  const handleOpenDialog = () => {
-    if (plannedEntries.length === 0) {
-      resetPlannedEntries();
-    }
-    setHasAttemptedSubmit(false);
-    setCategoryCreationTargetId(null);
-    setNewCategoryName('');
-    setIsAddExpenseDialogOpen(true);
-  };
-
-  const handleCancelDialog = () => {
-    setIsAddExpenseDialogOpen(false);
-    setCategoryCreationTargetId(null);
-    setNewCategoryName('');
-    setHasAttemptedSubmit(false);
-    resetPlannedEntries();
-  };
-
-  const handleCreateCategory = async () => {
-    if (!newCategoryName.trim()) return;
-    const category = await addCategory({
-      name: newCategoryName.trim(),
-      type: 'expense',
-      isCustom: true
-    });
-    setPlannedEntries((previous) =>
-      previous.map((entry) =>
-        categoryCreationTargetId && entry.id === categoryCreationTargetId
-          ? { ...entry, categoryId: category.id }
-          : entry
-      )
-    );
-    setNewCategoryName('');
-    setCategoryCreationTargetId(null);
-  };
-
-  const hasInvalidEntries = plannedEntries.some((entry) => !isEntryValid(entry));
-  const shouldShowValidationError =
-    hasAttemptedSubmit && (hasInvalidEntries || expenseCategories.length === 0);
-  const canRemoveRows = plannedEntries.length > 1;
-
-  const toggleCategory = (id: string) => {
-    setExpandedCategories((previous) => ({ ...previous, [id]: !previous[id] }));
-  };
-
-  function focusCategory(id: string, expandSelf = false) {
-    setFocusedCategoryId(id);
-    setExpandedCategories((previous) => {
-      const next = { ...previous } as Record<string, boolean>;
-      let currentParent = categoryParentMap.get(id) ?? null;
-      while (currentParent) {
-        next[currentParent] = true;
-        currentParent = categoryParentMap.get(currentParent) ?? null;
-      }
-      if (expandSelf) {
-        next[id] = true;
-      }
-      return next;
-    });
-  }
-
-  const handleSummaryScopeChange = (value: 'all' | string) => {
-    setSelectedCategoryId(value);
-    if (value !== 'all') {
-      focusCategory(value, true);
-    }
-  };
-
-  const handleMonthInputChange = (value: string) => {
-    if (!value) {
-      setSelectedMonth(defaultMonth);
-      return;
-    }
-    setSelectedMonth(value);
-  };
-
-  const handleYearInputChange = (value: string) => {
-    if (!value) {
-      setSelectedYear(defaultYear);
-      return;
-    }
-    const sanitised = value.replace(/[^0-9]/g, '').slice(0, 4);
-    setSelectedYear(sanitised || defaultYear);
-  };
-
-  const handleResetFilters = () => {
-    setNavigatorFilter('all');
-    setNavigatorView('category');
-    setCategorySearchTerm('');
-    setSelectedCategoryId('all');
-  };
-
-  const handleBudgetDraftChange = (value: string) => {
-    setBudgetDraft(value);
-    setBudgetError(null);
-    setBudgetFeedback(null);
-  };
-
-  const applyBudgetDraft = async (rawValue: string) => {
-    if (isSavingBudget) {
-      return;
-    }
-    if (!focusedCategory) {
-      setBudgetError('Choose a category in the navigator to set a baseline budget.');
-      return;
-    }
-    const trimmed = rawValue.trim();
-    const budgetKey = viewMode === 'monthly' ? 'monthly' : 'yearly';
-    const nextBudgets: Category['budgets'] = { ...(focusedCategory.budgets ?? {}) };
-    if (trimmed === '') {
-      delete nextBudgets[budgetKey];
-    } else {
-      const numericValue = Number(trimmed);
-      if (Number.isNaN(numericValue) || numericValue < 0) {
-        setBudgetError('Enter a valid amount to save.');
-        return;
-      }
-      nextBudgets[budgetKey] = numericValue;
-    }
-    setIsSavingBudget(true);
-    setBudgetError(null);
-    setBudgetFeedback(null);
-    try {
-      await updateCategory(focusedCategory.id, { budgets: nextBudgets });
-      setBudgetFeedback(trimmed === '' ? 'cleared' : 'saved');
-    } finally {
-      setIsSavingBudget(false);
-    }
-  };
-
-  const handleBudgetSubmit = async (event: FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
-    await applyBudgetDraft(budgetDraft);
-  };
-
-  const handleBudgetClear = async () => {
-    if (!focusedCategory) {
-      setBudgetError('Choose a category in the navigator to set a baseline budget.');
-      return;
-    }
-    setBudgetDraft('');
-    await applyBudgetDraft('');
-  };
-
-  const handleStartEdit = (detail: PlannedExpenseDetail) => {
-    setEditingItemId(detail.item.id);
-    const manualActual =
-      typeof detail.item.actualAmount === 'number' && !Number.isNaN(detail.item.actualAmount)
-        ? detail.item.actualAmount
-        : undefined;
-    const fallbackDate = formatISO(parseISO(detail.item.createdAt), { representation: 'date' });
-    const remainderSource =
-      typeof detail.item.remainderAmount === 'number' && !Number.isNaN(detail.item.remainderAmount)
-        ? detail.item.remainderAmount
-        : detail.remainder > 0
-        ? detail.remainder
-        : undefined;
-    setEditDraft({
-      categoryId: detail.item.categoryId,
-      plannedAmount: String(detail.item.plannedAmount),
-      actualAmount:
-        manualActual !== undefined
-          ? String(manualActual)
-          : detail.actual > 0
-          ? String(detail.actual)
-          : '',
-      remainderAmount:
-        remainderSource !== undefined && remainderSource > 0 ? String(remainderSource) : '',
-      dueDate: detail.item.dueDate ?? fallbackDate,
-      hasDueDate: Boolean(detail.item.dueDate),
-      priority: detail.item.priority ?? 'medium'
-    });
-  };
-
-  const handleCancelEdit = () => {
-    setEditingItemId(null);
-    setEditDraft({
-      categoryId: '',
-      plannedAmount: '',
-      actualAmount: '',
-      remainderAmount: '',
-      dueDate: '',
-      hasDueDate: true,
-      priority: 'medium'
-    });
-  };
-
-  const handleQuickActualChange = (id: string, value: string) => {
-    setQuickActualDrafts((previous) => ({ ...previous, [id]: value }));
-  };
-
-  const handleQuickActualSubmit = async (detail: PlannedExpenseDetail) => {
-    const rawValue = (quickActualDrafts[detail.item.id] ?? '').trim();
-    if (rawValue === '') {
-      return;
-    }
-    const numericValue = Number(rawValue);
-    if (Number.isNaN(numericValue) || numericValue < 0) {
-      return;
-    }
-    setQuickActualSavingId(detail.item.id);
-    try {
-      await updatePlannedExpense(detail.item.id, { actualAmount: numericValue });
-      setQuickActualDrafts((previous) => ({ ...previous, [detail.item.id]: '' }));
-    } finally {
-      setQuickActualSavingId(null);
-    }
-  };
-
-  const handleSaveEdit = async (detail: PlannedExpenseDetail) => {
-    const plannedValue = Number(editDraft.plannedAmount);
-    const trimmedActual = editDraft.actualAmount.trim();
-    const actualValue = trimmedActual === '' ? undefined : Number(trimmedActual);
-    const trimmedRemainder = editDraft.remainderAmount.trim();
-    const remainderValue = trimmedRemainder === '' ? null : Number(trimmedRemainder);
-    const requiresDueDate = editDraft.hasDueDate && editDraft.dueDate.trim() === '';
-    if (!editDraft.categoryId || Number.isNaN(plannedValue) || plannedValue < 0) {
-      return;
-    }
-    if (actualValue !== undefined && (Number.isNaN(actualValue) || actualValue < 0)) {
-      return;
-    }
-    if (remainderValue !== null && (Number.isNaN(remainderValue) || remainderValue < 0 || remainderValue > plannedValue)) {
-      return;
-    }
-    if (requiresDueDate) {
-      return;
-    }
-    setSavingItemId(detail.item.id);
-    try {
-      let nextActualValue = actualValue;
-      if (remainderValue !== null) {
-        nextActualValue = Math.max(plannedValue - remainderValue, 0);
-      }
-      await updatePlannedExpense(detail.item.id, {
-        categoryId: editDraft.categoryId,
-        plannedAmount: plannedValue,
-        actualAmount: nextActualValue,
-        remainderAmount: remainderValue,
-        priority: editDraft.priority,
-        dueDate: editDraft.hasDueDate
-          ? editDraft.dueDate || detail.item.dueDate || formatISO(new Date(), { representation: 'date' })
-          : null
-      });
-      setEditingItemId(null);
-      setEditDraft({
-        categoryId: '',
-        plannedAmount: '',
-        actualAmount: '',
-        remainderAmount: '',
-        dueDate: '',
-        hasDueDate: true,
-        priority: 'medium'
-      });
-    } finally {
-      setSavingItemId(null);
-    }
-  };
-
-  const handleFocusDetail = (detail: PlannedExpenseDetail) => {
-    setNavigatorFilter('all');
-    setCategorySearchTerm('');
-    setNavigatorView('category');
-    focusCategory(detail.item.categoryId, true);
-    handleStartEdit(detail);
-  };
-
-  const renderItemCard = (detail: PlannedExpenseDetail, depth: number) => {
-    const isEditing = editingItemId === detail.item.id;
-    const isSaving = savingItemId === detail.item.id;
-    const categoryName =
-      categoryLookup.get(detail.item.categoryId)?.name ??
-      categories.find((cat) => cat.id === detail.item.categoryId)?.name ??
-      'Uncategorised';
-    const dueDateLabel = detail.item.dueDate
-      ? new Date(detail.item.dueDate).toLocaleDateString('en-IN', {
-          month: 'short',
-          day: 'numeric'
-        })
-      : 'No due date';
-    const progressPercentRaw =
-      detail.item.plannedAmount <= 0
-        ? detail.actual > 0
-          ? 100
-          : 0
-        : (detail.actual / detail.item.plannedAmount) * 100;
-    const progressPercent = Number.isFinite(progressPercentRaw) ? progressPercentRaw : 0;
-    const progressWidth = Math.max(0, Math.min(100, progressPercent));
-    const progressColor = progressColorByStatus[detail.status];
-    const varianceLabel = detail.variance >= 0 ? 'Saved' : 'Overspent';
-    const statusToken = spendingBadgeStyles[detail.status];
-    const priorityToken = PRIORITY_TOKEN_STYLES[detail.priority ?? 'medium'];
-    const actualToneClass = statusToken.toneClass;
-    const actualBackgroundClass =
-      detail.status === 'over'
-        ? 'bg-danger/10'
-        : detail.status === 'under'
-        ? 'bg-success/10'
-        : 'bg-slate-950/80';
-    const remainderValue = detail.remainder;
-    const remainderLabel = remainderValue >= 0 ? 'Remaining' : 'Overspent';
-    const remainderDisplay = Math.abs(remainderValue);
-    const remainderToneClass = remainderValue >= 0 ? 'text-success' : 'text-danger';
-    const remainderBackgroundClass = remainderValue >= 0 ? 'bg-success/10' : 'bg-danger/10';
-    const isCurrentCategoryMissing =
-      isEditing && editDraft.categoryId && !categoryOptions.some((option) => option.id === editDraft.categoryId);
-    const parsedPlanned = Number(editDraft.plannedAmount);
-    const parsedActual = editDraft.actualAmount.trim() === '' ? undefined : Number(editDraft.actualAmount);
-    const parsedRemainder = editDraft.remainderAmount.trim() === '' ? null : Number(editDraft.remainderAmount);
-    const isRemainderProvided = editDraft.remainderAmount.trim() !== '';
-    const hasPlannedError = isEditing && (Number.isNaN(parsedPlanned) || parsedPlanned < 0);
-    const hasActualError =
-      isEditing &&
-      !isRemainderProvided &&
-      parsedActual !== undefined &&
-      (Number.isNaN(parsedActual) || parsedActual < 0);
-    const hasRemainderError =
-      isEditing &&
-      parsedRemainder !== null &&
-      (Number.isNaN(parsedRemainder) ||
-        parsedRemainder < 0 ||
-        (Number.isFinite(parsedPlanned) && parsedPlanned >= 0 && parsedRemainder > parsedPlanned));
-    const requiresDueDate = isEditing && editDraft.hasDueDate && editDraft.dueDate.trim() === '';
-    const isSaveDisabled =
-      !isEditing ||
-      !editDraft.categoryId ||
-      editDraft.plannedAmount.trim() === '' ||
-      hasPlannedError ||
-      hasActualError ||
-      hasRemainderError ||
-      requiresDueDate ||
-      isSaving;
-    const quickActualDraft = quickActualDrafts[detail.item.id] ?? '';
-    const quickActualTrimmed = quickActualDraft.trim();
-    const quickActualValue = quickActualTrimmed === '' ? undefined : Number(quickActualTrimmed);
-    const hasQuickActualError =
-      quickActualValue !== undefined && (Number.isNaN(quickActualValue) || quickActualValue < 0);
-    const isQuickSaving = quickActualSavingId === detail.item.id;
-    const quickPlaceholder =
-      typeof detail.item.actualAmount === 'number' && !Number.isNaN(detail.item.actualAmount)
-        ? String(detail.item.actualAmount)
-        : detail.actual > 0
-        ? String(detail.actual)
-        : '';
-    const infoMessage =
-      typeof detail.item.actualAmount === 'number' && !Number.isNaN(detail.item.actualAmount)
-        ? `Manual spend recorded: ${formatCurrency(detail.actual)}.`
-        : detail.match
-        ? `Matched with ${detail.match.description} on ${new Date(detail.match.date).toLocaleDateString('en-IN')}`
-        : 'No matching transaction yet — update spent once the payment is made.';
-    const remainderColor = detail.variance >= 0 ? 'text-success' : 'text-danger';
-
-    const handleSubmitQuickActual = (event: FormEvent<HTMLFormElement>) => {
-      event.preventDefault();
-      if (!hasQuickActualError && quickActualValue !== undefined && !isQuickSaving) {
-        void handleQuickActualSubmit(detail);
-      }
-    };
-
-    const rowClass =
-      'group/row grid grid-cols-[minmax(0,2.6fr)_minmax(110px,0.9fr)_minmax(120px,0.9fr)_minmax(120px,0.9fr)_minmax(120px,0.9fr)_minmax(220px,1fr)] items-start gap-4 border-t border-slate-800/60 px-4 py-3 text-[11px] sm:text-xs transition';
-    const rowStateClass = isEditing ? 'bg-slate-950/60 ring-1 ring-inset ring-accent/40' : 'bg-slate-950/25 hover:bg-slate-900/55';
-
-    return (
-      <div key={detail.item.id} className={`${rowClass} ${rowStateClass}`}>
-        <div className="min-w-0 space-y-2" style={{ paddingLeft: depth * 20 }}>
-          <div className="flex flex-wrap items-center justify-between gap-2">
-            <div className="flex min-w-0 flex-wrap items-center gap-2">
-              <span className="truncate text-sm font-semibold text-slate-100">{detail.item.name}</span>
-              <span className={`rounded-full px-2 py-0.5 text-[10px] font-semibold ${statusToken.badgeClass}`}>
-                {statusToken.label}
-              </span>
-            </div>
-            <span className="text-[10px] text-slate-500" title={categoryName}>
-              {categoryName}
-            </span>
-          </div>
-          <p className="truncate text-[10px] text-slate-500" title={infoMessage}>
-            {infoMessage}
-          </p>
-          <div className="flex items-center gap-2">
-            <div className="h-1.5 w-full flex-1 overflow-hidden rounded-full bg-slate-800">
-              <div className="h-full rounded-full" style={{ width: `${progressWidth}%`, backgroundColor: progressColor }} />
-            </div>
-            <p className="text-[10px] text-slate-500">{infoMessage}</p>
-            {isEditing && (
-              <div className="pt-1">
-                <label className="text-[10px] uppercase text-slate-500">Category</label>
-                <select
-                  className="mt-1 w-full rounded-md border border-slate-700 bg-slate-950/80 px-3 py-1.5 text-xs text-slate-100 focus:border-accent focus:outline-none"
-                  value={editDraft.categoryId}
-                  onChange={(event) => setEditDraft((prev) => ({ ...prev, categoryId: event.target.value }))}
-                  disabled={isSaving}
-                >
-                  {isCurrentCategoryMissing && (
-                    <option value={detail.item.categoryId}>
-                      {categories.find((cat) => cat.id === detail.item.categoryId)?.name ?? 'Uncategorised'}
-                    </option>
-                  )}
-                  {categoryOptions.map((option) => (
-                    <option key={option.id} value={option.id}>
-                      {option.label}
-                    </option>
-                  ))}
-                </select>
-                {isCurrentCategoryMissing && (
-                  <p className="mt-1 text-[10px] text-warning">
-                    The original category is no longer available. Pick another one before saving.
-                  </p>
-                )}
-              </div>
-            )}
-          </div>
-        </div>
-        <div className="flex flex-col justify-center gap-1 text-xs text-slate-300">
-          <span className="text-sm font-semibold text-slate-100">{dueDateLabel}</span>
-          <div className="flex flex-wrap items-center gap-1 text-[10px] text-slate-500">{statusBadge(detail.item.status)}</div>
-        </div>
-        <div className="text-right">
-          {isEditing ? (
-            <div className="space-y-1">
-              <input
-                type="number"
-                min={0}
-                className={`w-full rounded-md border bg-slate-950/80 px-3 py-1.5 text-sm text-slate-100 focus:border-accent focus:outline-none ${
-                  hasPlannedError ? 'border-danger text-danger focus:border-danger' : 'border-slate-700'
-                }`}
-                value={editDraft.plannedAmount}
-                onChange={(event) => setEditDraft((prev) => ({ ...prev, plannedAmount: event.target.value }))}
-                disabled={isSaving}
-              />
-              {hasPlannedError && <p className="text-[10px] text-danger">Enter a valid planned amount.</p>}
-            </div>
-          ) : (
-            <div className="space-y-1">
-              <div className="text-sm font-semibold text-warning">{formatCurrency(detail.item.plannedAmount)}</div>
-              <div className="text-[10px] text-slate-500">Planned</div>
-            </div>
-          )}
-        </div>
-        <div className="text-right">
-          {isEditing ? (
-            <div className="space-y-1">
-              <input
-                type="number"
-                min={0}
-                placeholder="Auto from transactions"
-                className={`w-full rounded-md border bg-slate-950/80 px-3 py-1.5 text-sm text-slate-100 focus:border-accent focus:outline-none ${
-                  hasActualError ? 'border-danger text-danger focus:border-danger' : 'border-slate-700'
-                }`}
-                value={editDraft.actualAmount}
-                onChange={(event) => setEditDraft((prev) => ({ ...prev, actualAmount: event.target.value }))}
-                disabled={isRemainderProvided || isSaving}
-              />
-              {hasActualError && <p className="text-[10px] text-danger">Enter a valid spent amount.</p>}
-            </div>
-          ) : (
-            <div className={`space-y-1 rounded-md border border-slate-800/70 px-3 py-1.5 text-right ${actualBackgroundClass}`}>
-              <div className={`text-sm font-semibold ${actualToneClass}`}>{formatCurrency(detail.actual)}</div>
-              <div className="text-[10px] text-slate-500">Spent</div>
-            </div>
-          )}
-        </div>
-        <div className="text-right">
-          <div className={`text-sm font-semibold ${remainderColor}`}>{formatCurrency(detail.variance)}</div>
-          <div className="text-[10px] text-slate-500">{varianceLabel}</div>
-        </div>
-        <div className="flex flex-col items-end gap-2 text-right">
-          {!isEditing && (
-            <form className="flex w-full items-center justify-end gap-2" onSubmit={handleSubmitQuickActual}>
-              <input
-                type="number"
-                min={0}
-                value={quickActualDraft}
-                onChange={(event) => handleQuickActualChange(detail.item.id, event.target.value)}
-                placeholder={quickPlaceholder || 'Spent'}
-                className={`w-24 rounded-md border bg-slate-950/80 px-2 py-1 text-xs text-slate-100 focus:border-accent focus:outline-none ${
-                  hasQuickActualError ? 'border-danger text-danger focus:border-danger' : 'border-slate-700'
-                }`}
-              />
-              <button
-                type="submit"
-                disabled={hasQuickActualError || quickActualValue === undefined || isQuickSaving}
-                className="rounded-md bg-accent px-2 py-1 text-[11px] font-semibold text-slate-900 transition hover:bg-accent/90 disabled:cursor-not-allowed disabled:opacity-60"
-              >
-                {isQuickSaving ? 'Saving…' : 'Save'}
-              </button>
-            </form>
-          )}
-          {hasQuickActualError && !isEditing && (
-            <p className="text-[10px] text-danger">Enter a valid amount to save.</p>
-          )}
-          <div className="flex flex-wrap justify-end gap-1 text-[10px]">
-            <button
-              type="button"
-              className="rounded-full bg-success/15 px-2 py-1 font-semibold text-success hover:bg-success/25"
-              onClick={() => updatePlannedExpense(detail.item.id, { status: 'purchased' })}
-            >
-              Purchased
-            </button>
-            <button
-              type="button"
-              className="rounded-full bg-slate-800 px-2 py-1 text-slate-300 hover:bg-slate-700"
-              onClick={() => updatePlannedExpense(detail.item.id, { status: 'cancelled' })}
-            >
-              Cancel
-            </button>
-            <button
-              type="button"
-              className="rounded-full bg-sky-500/15 px-2 py-1 font-semibold text-sky-300 hover:bg-sky-500/25"
-              onClick={() => updatePlannedExpense(detail.item.id, { status: 'reconciled' })}
-              disabled={detail.item.status === 'reconciled'}
-            >
-              Reconcile
-            </button>
-            <button
-              type="button"
-              className="rounded-full bg-danger/15 px-2 py-1 font-semibold text-danger hover:bg-danger/25"
-              onClick={() => deletePlannedExpense(detail.item.id)}
-            >
-              Delete
-            </button>
-          </div>
-          {isEditing ? (
-            <div className="flex flex-wrap justify-end gap-2">
-              <button
-                type="button"
-                onClick={() => void handleSaveEdit(detail)}
-                disabled={isSaveDisabled}
-                className="rounded-md bg-success px-3 py-1 text-[11px] font-semibold text-slate-900 transition hover:bg-emerald-400 disabled:cursor-not-allowed disabled:opacity-60"
-              >
-                {isSaving ? 'Saving…' : 'Save changes'}
-              </button>
-              <button
-                type="button"
-                onClick={handleCancelEdit}
-                disabled={isSaving}
-                className="rounded-md border border-slate-700 px-3 py-1 text-[11px] font-semibold text-slate-300 hover:border-accent hover:text-accent disabled:cursor-not-allowed disabled:opacity-60"
-              >
-                Cancel
-              </button>
-            </div>
-          ) : (
-            <button
-              type="button"
-              onClick={() => handleStartEdit(detail)}
-              className="text-[11px] font-semibold text-accent hover:text-accent/80"
-            >
-              Edit details
-            </button>
-          )}
-        </div>
-      </div>
-    );
-  };
-
-  const renderCategorySection = (category: CategoryNode, depth = 0): JSX.Element | null => {
-    const summary = categorySummaries.get(category.id) ?? {
-      planned: 0,
-      actual: 0,
-      variance: 0,
-      itemCount: 0,
-      remainder: 0
-    };
-    const directItems = itemsByCategory.get(category.id) ?? [];
-    const categoryStatus: PlannedExpenseSpendingHealth =
-      summary.actual === 0 ? 'not-spent' : summary.variance >= 0 ? 'under' : 'over';
-    const statusToken = spendingBadgeStyles[categoryStatus];
-    const descendantIds = expenseDescendantsMap.get(category.id) ?? new Set<string>([category.id]);
-    const remainderValue = summary.remainder;
-    const remainderClass = remainderValue >= 0 ? 'text-success' : 'text-danger';
-    const remainderLabel = remainderValue >= 0 ? 'Remaining' : 'Overspent';
-    const remainderDescriptor = summary.actual === 0 && remainderValue >= 0 ? 'Awaiting spend' : remainderLabel;
-    const nextDueDetail = plannedExpenseDetails.reduce<PlannedExpenseDetail | null>((closest, detail) => {
-      if (!descendantIds.has(detail.item.categoryId) || !detail.item.dueDate) {
-        return closest;
-      }
-      if (!closest || !closest.item.dueDate) return detail;
-      const currentTime = new Date(detail.item.dueDate).getTime();
-      const closestTime = new Date(closest.item.dueDate).getTime();
-      return currentTime < closestTime ? detail : closest;
-    }, null);
-    const nextDueLabel =
-      nextDueDetail && nextDueDetail.item.dueDate
-        ? new Date(nextDueDetail.item.dueDate).toLocaleDateString('en-IN', {
-            month: 'short',
-            day: 'numeric'
-          })
-        : null;
-    const matchesCategorySearch =
-      normalisedSearchTerm !== '' && category.name.toLowerCase().includes(normalisedSearchTerm);
-    const visibleDirectItems = directItems.filter((detail) => {
-      const matchesStatus = navigatorFilter === 'all' || detail.status === navigatorFilter;
-      const matchesName =
-        normalisedSearchTerm === '' ||
-        detail.item.name.toLowerCase().includes(normalisedSearchTerm) ||
-        matchesCategorySearch;
-      return matchesStatus && matchesName;
-    });
-    const childSections = category.children
-      .map((child) => renderCategorySection(child, depth + 1))
-      .filter((child): child is JSX.Element => child !== null);
-    const hasVisibleItems = visibleDirectItems.length > 0;
-    const hasVisibleChildren = childSections.length > 0;
-    const canExpand = hasVisibleItems || hasVisibleChildren;
-    const matchesStatusForCategory = navigatorFilter === 'all' || categoryStatus === navigatorFilter;
-
-    if (!matchesStatusForCategory && !hasVisibleItems && !hasVisibleChildren && !matchesCategorySearch) {
-      return null;
-    }
-
-    // Always render expense categories so they can be focused for baseline adjustments,
-    // even if they don't yet have planned items in the current period.
-
-    const isFocused = focusedCategoryId === category.id;
-    const shouldAutoExpand =
-      normalisedSearchTerm !== '' && (matchesCategorySearch || hasVisibleItems || hasVisibleChildren);
-    const isExpanded = canExpand && (shouldAutoExpand || Boolean(expandedCategories[category.id]));
-    const focusClass = isFocused ? 'bg-slate-900/60 ring-1 ring-inset ring-accent/40' : 'bg-slate-950/30 hover:bg-slate-900/50';
-    const dimClass =
-      normalisedSearchTerm !== '' && !matchesCategorySearch && !hasVisibleItems && !hasVisibleChildren
-        ? 'opacity-70'
-        : '';
-    const handleToggle = () => {
-      focusCategory(category.id);
-      if (canExpand) {
-        toggleCategory(category.id);
-      }
-    };
-    const indentation = depth * 20;
-
-    return (
-      <div key={category.id} className={dimClass}>
-        <div
-          onClick={() => focusCategory(category.id, true)}
-          className={`grid cursor-pointer grid-cols-[minmax(0,2.6fr)_minmax(110px,0.9fr)_minmax(120px,0.9fr)_minmax(120px,0.9fr)_minmax(120px,0.9fr)_minmax(220px,1fr)] items-center gap-4 border-t border-slate-800/70 px-4 py-3 text-[11px] sm:text-xs transition ${focusClass}`}
-        >
-          <div className="flex min-w-0 items-center gap-3" style={{ paddingLeft: indentation }}>
-            <button
-              type="button"
-              onClick={(event) => {
-                event.stopPropagation();
-                handleToggle();
-              }}
-              aria-expanded={Boolean(isExpanded)}
-              aria-disabled={!canExpand}
-              className={`flex h-6 w-6 items-center justify-center rounded-md border border-slate-700 bg-slate-950/60 text-slate-400 transition ${
-                canExpand ? 'hover:border-accent hover:text-accent' : 'cursor-default opacity-40'
-              }`}
-            >
-              <span aria-hidden className={`text-sm transition-transform ${isExpanded ? 'rotate-90' : ''}`}>
-                {canExpand ? '▸' : '•'}
-              </span>
-            </button>
-            <div className="min-w-0 space-y-1">
-              <div className="flex flex-wrap items-center gap-2">
-                <p className="truncate text-sm font-semibold text-slate-100">{category.name}</p>
-                <span className={`rounded-full px-2 py-0.5 text-[10px] font-semibold ${statusToken.badgeClass}`}>
-                  {statusToken.label}
-                </span>
-              </div>
-              <div className="flex flex-wrap items-center gap-2 text-[10px] text-slate-500">
-                <span>
-                  {summary.itemCount} planned item{summary.itemCount === 1 ? '' : 's'}
-                </span>
-                {nextDueLabel && <span>Next due {nextDueLabel}</span>}
-              </div>
-            </div>
-          </div>
-          <div className="space-y-1 text-xs text-slate-300">
-            <span className="text-sm font-semibold text-slate-100">{nextDueLabel ?? '—'}</span>
-            <span className="text-[10px] text-slate-500">
-              {nextDueLabel ? 'Earliest due' : 'No due dates'}
-            </span>
-          </div>
-          <div className="text-right text-sm font-semibold text-warning">{formatCurrency(summary.planned)}</div>
-          <div className="text-right text-sm font-semibold text-slate-200">{formatCurrency(summary.actual)}</div>
-          <div className="text-right">
-            <div className={`text-sm font-semibold ${remainderClass}`}>{formatCurrency(summary.variance)}</div>
-            <div className="text-[10px] text-slate-500">{remainderDescriptor}</div>
-          </div>
-          <div className="flex flex-wrap items-center justify-end gap-2 text-[10px] text-slate-400">
-            {nextDueLabel && (
-              <span className="rounded-full bg-slate-800/70 px-2 py-0.5 text-slate-300">Next {nextDueLabel}</span>
-            )}
-            {canExpand && <span>{isExpanded ? 'Collapse' : 'Expand'}</span>}
-          </div>
-        </div>
-        {isExpanded && (
-          <div className="bg-slate-950/15">
-            {visibleDirectItems.length > 0 && visibleDirectItems.map((detail) => renderItemCard(detail, depth + 1))}
-            {childSections}
-          </div>
-        )}
-      </div>
-    );
-  };
-
-  const renderedCategorySections = expenseCategoryTree
-    .map((category) => renderCategorySection(category))
-    .filter((section): section is JSX.Element => section !== null);
-  const summaryPeriodLabel = viewMode === 'monthly' ? 'Monthly' : 'Yearly';
-  const summaryPeriodDescriptor = viewMode === 'monthly' ? 'month' : 'year';
-  const overallStatusToken = spendingBadgeStyles[overallSummary.status];
-  const selectedCategoryLabel =
-    selectedCategoryId === 'all'
-      ? 'All categories'
-      : categoryLookup.get(selectedCategoryId)?.name ?? 'All categories';
-  const selectedCategoryVariance = totalsForSelected.totalPlanned - totalsForSelected.actualTotal;
-  const selectedCategoryStatus: PlannedExpenseSpendingHealth =
-    totalsForSelected.actualTotal === 0
-      ? 'not-spent'
-      : selectedCategoryVariance >= 0
-      ? 'under'
-      : 'over';
-  const selectedStatusToken = spendingBadgeStyles[selectedCategoryStatus];
-  const baselineLabel = viewMode === 'monthly' ? 'Monthly baseline (₹)' : 'Yearly baseline (₹)';
-  const hasNavigatorResults = renderedCategorySections.length > 0 || visibleUncategorisedDetails.length > 0;
   return (
     <div className="space-y-6">
-      {isAddExpenseDialogOpen && (
-        <div className="fixed inset-0 z-50 flex items-start justify-center overflow-y-auto bg-slate-950/80 px-4 py-6 backdrop-blur">
-          <div className="w-full max-w-4xl rounded-2xl border border-slate-800 bg-slate-900/95 p-6 shadow-2xl">
-            <form onSubmit={handleSubmitPlannedExpenses} className="space-y-6">
-              <div className="flex flex-col gap-4 sm:flex-row sm:items-start sm:justify-between">
-                <div>
-                  <h3 className="text-xl font-semibold text-slate-100">Add planned expenses</h3>
-                  <p className="text-sm text-slate-400">
-                    Capture multiple planned expenses at once and assign them to categories.
-                  </p>
-                </div>
-                <button
-                  type="button"
-                  onClick={handleCancelDialog}
-                  className="self-start rounded-lg border border-slate-700 px-3 py-1 text-xs font-semibold uppercase tracking-wide text-slate-300 transition hover:border-slate-500"
-                >
-                  Close
-                </button>
-              </div>
+      <PlannedExpensesDialog dialog={dialog} utils={utils} />
 
-              <div className="overflow-x-auto">
-                <table className="min-w-full divide-y divide-slate-800 text-sm">
-                  <thead>
-                    <tr className="text-left text-xs uppercase tracking-wide text-slate-500">
-                      <th className="px-3 py-2 font-semibold">Name</th>
-                      <th className="px-3 py-2 font-semibold">Amount (₹)</th>
-                      <th className="px-3 py-2 font-semibold">Due date</th>
-                      <th className="px-3 py-2 font-semibold">Priority</th>
-                      <th className="px-3 py-2 font-semibold">Category</th>
-                      <th className="px-3 py-2 font-semibold">Actions</th>
-                    </tr>
-                  </thead>
-                  <tbody className="divide-y divide-slate-800">
-                    {plannedEntries.map((entry, index) => {
-                      const isCreatingForRow = categoryCreationTargetId === entry.id;
-                      return (
-                        <tr key={entry.id} className="align-top">
-                          <td className="px-3 py-2">
-                            <input
-                              className="w-full rounded-lg border border-slate-800 bg-slate-950 px-3 py-2 text-sm"
-                              placeholder="e.g. School fees"
-                              value={entry.name}
-                              onChange={(event) => handleEntryChange(entry.id, { name: event.target.value })}
-                            />
-                          </td>
-                          <td className="px-3 py-2">
-                            <input
-                              type="number"
-                              min={0}
-                              className="w-full rounded-lg border border-slate-800 bg-slate-950 px-3 py-2 text-sm"
-                              placeholder="0"
-                              value={entry.amount}
-                              onChange={(event) => handleEntryChange(entry.id, { amount: event.target.value })}
-                            />
-                          </td>
-                          <td className="px-3 py-2">
-                            <div className="flex flex-col gap-2">
-                              <input
-                                type="date"
-                                className="w-full rounded-lg border border-slate-800 bg-slate-950 px-3 py-2 text-sm disabled:opacity-60"
-                                value={entry.dueDate}
-                                onChange={(event) =>
-                                  handleEntryChange(entry.id, { dueDate: event.target.value })
-                                }
-                                disabled={!entry.hasDueDate}
-                              />
-                              <label className="flex items-center gap-2 text-xs text-slate-400">
-                                <input
-                                  type="checkbox"
-                                  checked={!entry.hasDueDate}
-                                  onChange={(event) => {
-                                    const noDueDate = event.target.checked;
-                                    handleEntryChange(entry.id, {
-                                      hasDueDate: !noDueDate,
-                                      ...(noDueDate
-                                        ? {}
-                                        : { dueDate: entry.dueDate || resolveDefaultDueDate() })
-                                    });
-                                  }}
-                                />
-                                <span>No due date</span>
-                              </label>
-                            </div>
-                          </td>
-                          <td className="px-3 py-2">
-                            <select
-                              className="w-full rounded-lg border border-slate-800 bg-slate-950 px-3 py-2 text-sm"
-                              value={entry.priority}
-                              onChange={(event) =>
-                                handleEntryChange(entry.id, {
-                                  priority: event.target.value as PlannedExpenseItem['priority']
-                                })
-                              }
-                            >
-                              {PRIORITY_OPTIONS.map((option) => (
-                                <option key={option.value} value={option.value}>
-                                  {option.label}
-                                </option>
-                              ))}
-                            </select>
-                          </td>
-                          <td className="px-3 py-2">
-                            <div className="flex flex-col gap-2">
-                              <select
-                                className="w-full rounded-lg border border-slate-800 bg-slate-950 px-3 py-2 text-sm"
-                                value={entry.categoryId}
-                                onChange={(event) => handleEntryChange(entry.id, { categoryId: event.target.value })}
-                                disabled={expenseCategories.length === 0}
-                              >
-                                <option value="" disabled>
-                                  {expenseCategories.length === 0 ? 'No categories available' : 'Select category'}
-                                </option>
-                                {expenseCategories.map((category) => (
-                                  <option key={category.id} value={category.id}>
-                                    {category.name}
-                                  </option>
-                                ))}
-                              </select>
-                              <button
-                                type="button"
-                                onClick={() => handleToggleCategoryCreation(entry.id)}
-                                className="self-start text-xs font-semibold text-accent"
-                              >
-                                {isCreatingForRow ? 'Cancel new category' : 'New category'}
-                              </button>
-                            </div>
-                          </td>
-                          <td className="px-3 py-2">
-                            <button
-                              type="button"
-                              onClick={() => handleRemoveEntryRow(entry.id)}
-                              className="rounded-lg border border-slate-800 px-3 py-2 text-xs font-semibold text-slate-300 transition hover:border-slate-600 hover:text-slate-100 disabled:opacity-40"
-                              disabled={!canRemoveRows}
-                            >
-                              Remove
-                            </button>
-                          </td>
-                        </tr>
-                      );
-                    })}
-                  </tbody>
-                </table>
-              </div>
+      <SummaryHeaderControls
+        period={period}
+        onOpenDialog={dialog.open}
+        onExpandAll={categories.expandAllCategories}
+        onCollapseAll={categories.collapseAllCategories}
+      />
 
-              {categoryCreationTargetId && (
-                <div className="rounded-xl border border-slate-800 bg-slate-950/70 p-4">
-                  <p className="text-sm font-semibold text-slate-200">Create a new category</p>
-                  <p className="text-xs text-slate-500">
-                    The new category will automatically be assigned to the selected planned expense row.
-                  </p>
-                  <div className="mt-3 flex flex-col gap-2 sm:flex-row">
-                    <input
-                      className="w-full rounded-lg border border-slate-800 bg-slate-950 px-3 py-2 text-sm"
-                      placeholder="Category name"
-                      value={newCategoryName}
-                      onChange={(event) => setNewCategoryName(event.target.value)}
-                    />
-                    <div className="flex gap-2">
-                      <button
-                        type="button"
-                        onClick={handleCreateCategory}
-                        className="rounded-lg bg-accent px-3 py-2 text-xs font-semibold text-slate-900 disabled:opacity-50"
-                        disabled={!newCategoryName.trim()}
-                      >
-                        Save
-                      </button>
-                      <button
-                        type="button"
-                        onClick={() => {
-                          setCategoryCreationTargetId(null);
-                          setNewCategoryName('');
-                        }}
-                        className="rounded-lg border border-slate-700 px-3 py-2 text-xs font-semibold text-slate-300 transition hover:border-slate-500"
-                      >
-                        Cancel
-                      </button>
-                    </div>
-                  </div>
-                </div>
-              )}
+      <SummaryGrid overview={overview} categories={categories} inspector={inspector} utils={utils} />
 
-              {shouldShowValidationError && (
-                <p className="text-sm text-danger">
-                  Please complete every row or remove the ones you do not need before saving.
-                </p>
-              )}
-              {!expenseCategories.length && (
-                <p className="text-sm text-warning">
-                  Add an expense category first to start planning your expenses.
-                </p>
-              )}
+      <NavigatorFilters categories={categories} />
 
-              <div className="flex flex-col gap-3 sm:flex-row sm:items-center sm:justify-between">
-                <button
-                  type="button"
-                  onClick={handleAddEntryRow}
-                  className="inline-flex items-center justify-center rounded-lg border border-dashed border-slate-700 px-4 py-2 text-sm font-semibold text-slate-200 transition hover:border-slate-500"
-                >
-                  + Add another row
-                </button>
-                <div className="flex flex-col-reverse gap-2 sm:flex-row sm:items-center">
-                  <button
-                    type="button"
-                    onClick={handleCancelDialog}
-                    className="rounded-lg border border-slate-700 px-4 py-2 text-sm font-semibold text-slate-300 transition hover:border-slate-500"
-                  >
-                    Cancel
-                  </button>
-                  <button
-                    type="submit"
-                    className="rounded-lg bg-success px-4 py-2 text-sm font-semibold text-slate-900 transition hover:bg-emerald-400 disabled:opacity-50"
-                    disabled={
-                      isSubmittingPlannedExpenses ||
-                      hasInvalidEntries ||
-                      expenseCategories.length === 0
-                    }
-                  >
-                    {isSubmittingPlannedExpenses ? 'Saving...' : 'Save planned expenses'}
-                  </button>
-                </div>
-              </div>
-            </form>
-          </div>
+      <div className="flex flex-col gap-6 lg:flex-row lg:items-start lg:justify-between">
+        <div className="flex-1 space-y-4">
+          <CategoryNavigator categories={categories} editing={editing} utils={utils} />
         </div>
-      )}
-
-      <header>
-        <h2 className="text-2xl font-semibold">Smart Budgeting & Planned Expenses</h2>
-        <p className="text-sm text-slate-400">
-          Build proactive shopping lists, leverage AI suggestions for categories, and reconcile budgets with actual spends.
-        </p>
-      </header>
-
-      <section className="rounded-2xl border border-slate-800 bg-slate-900/60 p-4 sm:p-6">
-        <div className="space-y-6">
-          <div className="flex flex-col gap-4 xl:flex-row xl:items-center xl:justify-between">
-            <div className="flex flex-col gap-3">
-              <div className="flex flex-wrap items-center gap-3">
-                <div className="inline-flex rounded-lg border border-slate-700 bg-slate-950/70 p-1 text-xs font-semibold text-slate-300">
-                  <button
-                    type="button"
-                    onClick={() => handleViewModeChange('monthly')}
-                    className={`rounded-md px-3 py-1 transition ${
-                      viewMode === 'monthly'
-                        ? 'bg-accent text-slate-900'
-                        : 'hover:bg-slate-800 hover:text-slate-100'
-                    }`}
-                  >
-                    Monthly
-                  </button>
-                  <button
-                    type="button"
-                    onClick={() => handleViewModeChange('yearly')}
-                    className={`rounded-md px-3 py-1 transition ${
-                      viewMode === 'yearly'
-                        ? 'bg-accent text-slate-900'
-                        : 'hover:bg-slate-800 hover:text-slate-100'
-                    }`}
-                  >
-                    Yearly
-                  </button>
-                </div>
-                <div className="flex flex-wrap items-center gap-2 rounded-lg border border-slate-700 bg-slate-950/70 px-3 py-2 text-sm text-slate-200">
-                  <button
-                    type="button"
-                    onClick={goToPreviousPeriod}
-                    className="rounded-md border border-slate-700 px-2 py-1 text-xs font-semibold text-slate-300 transition hover:border-slate-500 hover:text-accent"
-                    aria-label="Previous period"
-                  >
-                    ‹
-                  </button>
-                  <span className="text-sm font-semibold text-slate-100">{periodLabel}</span>
-                  <button
-                    type="button"
-                    onClick={goToNextPeriod}
-                    className="rounded-md border border-slate-700 px-2 py-1 text-xs font-semibold text-slate-300 transition hover:border-slate-500 hover:text-accent"
-                    aria-label="Next period"
-                  >
-                    ›
-                  </button>
-                </div>
-                {viewMode === 'monthly' ? (
-                  <input
-                    type="month"
-                    value={selectedMonth}
-                    onChange={(event) => handleMonthInputChange(event.target.value)}
-                    className="rounded-lg border border-slate-700 bg-slate-950/70 px-3 py-2 text-sm text-slate-100 focus:border-accent focus:outline-none"
-                  />
-                ) : (
-                  <input
-                    type="text"
-                    inputMode="numeric"
-                    pattern="[0-9]*"
-                    maxLength={4}
-                    value={selectedYear}
-                    onChange={(event) => handleYearInputChange(event.target.value)}
-                    className="w-24 rounded-lg border border-slate-700 bg-slate-950/70 px-3 py-2 text-sm text-slate-100 focus:border-accent focus:outline-none"
-                  />
-                )}
-              </div>
-              <p className="text-xs text-slate-500">
-                Track planned versus actual spending for this {summaryPeriodDescriptor} and fine-tune category guardrails.
-              </p>
-            </div>
-            <div className="flex flex-wrap items-center gap-2">
-              <button
-                type="button"
-                onClick={handleOpenDialog}
-                className="inline-flex items-center gap-2 rounded-lg bg-accent px-4 py-2 text-sm font-semibold text-slate-900 transition hover:bg-accent/90"
-              >
-                + Plan expenses
-              </button>
-              <button
-                type="button"
-                onClick={expandAllCategories}
-                className="rounded-lg border border-slate-700 px-3 py-2 text-xs font-semibold text-slate-300 transition hover:border-slate-500"
-              >
-                Expand all
-              </button>
-              <button
-                type="button"
-                onClick={collapseAllCategories}
-                className="rounded-lg border border-slate-700 px-3 py-2 text-xs font-semibold text-slate-300 transition hover:border-slate-500"
-              >
-                Collapse all
-              </button>
-            </div>
-          </div>
-
-          <div className="grid gap-4 md:grid-cols-2 xl:grid-cols-4">
-            <div className="rounded-xl border border-slate-800 bg-slate-950/60 p-4">
-              <div className="flex items-start justify-between gap-3">
-                <div>
-                  <p className="text-xs font-semibold uppercase tracking-wide text-slate-400">{summaryPeriodLabel} overview</p>
-                  <p className="mt-1 text-lg font-semibold text-slate-100">{periodLabel}</p>
-                  <p className="mt-2 text-xs text-slate-400">
-                    Planned {formatCurrency(totalsForAll.totalPlanned)} · Spent {formatCurrency(totalsForAll.actualTotal)}
-                  </p>
-                </div>
-                <span className={`rounded-full px-2 py-1 text-[10px] font-semibold ${overallStatusToken.badgeClass}`}>
-                  {overallStatusToken.label}
-                </span>
-              </div>
-              <dl className="mt-4 grid grid-cols-3 gap-3 text-[11px] text-slate-400">
-                <div>
-                  <dt className="uppercase tracking-wide">Planned</dt>
-                  <dd className="mt-1 text-sm font-semibold text-warning">{formatCurrency(totalsForAll.plannedFromItems)}</dd>
-                </div>
-                <div>
-                  <dt className="uppercase tracking-wide">Budgets</dt>
-                  <dd className="mt-1 text-sm font-semibold text-slate-200">{formatCurrency(totalsForAll.budgetTotal)}</dd>
-                </div>
-                <div>
-                  <dt className="uppercase tracking-wide">Actual</dt>
-                  <dd className="mt-1 text-sm font-semibold text-slate-100">{formatCurrency(totalsForAll.actualTotal)}</dd>
-                </div>
-              </dl>
-              <div className="mt-4">
-                <div className="flex items-center justify-between text-[10px] text-slate-500">
-                  <span>Utilisation</span>
-                  <span>{overallUtilisationPercent}%</span>
-                </div>
-                <div className="mt-1 h-2 w-full overflow-hidden rounded-full bg-slate-800">
-                  <div className="h-full rounded-full bg-accent" style={{ width: `${overallUtilisationWidth}%` }} />
-                </div>
-              </div>
-            </div>
-
-            <div className="rounded-xl border border-slate-800 bg-slate-950/60 p-4">
-              <div className="flex items-start justify-between gap-3">
-                <div>
-                  <p className="text-xs font-semibold uppercase tracking-wide text-slate-400">Summary scope</p>
-                  <p className="mt-1 text-lg font-semibold text-slate-100">{selectedCategoryLabel}</p>
-                  <p className="mt-2 text-xs text-slate-400">
-                    Planned {formatCurrency(totalsForSelected.totalPlanned)} · Spent {formatCurrency(totalsForSelected.actualTotal)}
-                  </p>
-                </div>
-                <span className={`rounded-full px-2 py-1 text-[10px] font-semibold ${selectedStatusToken.badgeClass}`}>
-                  {selectedStatusToken.label}
-                </span>
-              </div>
-              <dl className="mt-4 grid grid-cols-3 gap-3 text-[11px] text-slate-400">
-                <div>
-                  <dt className="uppercase tracking-wide">Planned</dt>
-                  <dd className="mt-1 text-sm font-semibold text-warning">{formatCurrency(totalsForSelected.plannedFromItems)}</dd>
-                </div>
-                <div>
-                  <dt className="uppercase tracking-wide">Budgets</dt>
-                  <dd className="mt-1 text-sm font-semibold text-slate-200">{formatCurrency(totalsForSelected.budgetTotal)}</dd>
-                </div>
-                <div>
-                  <dt className="uppercase tracking-wide">Variance</dt>
-                  <dd className={`mt-1 text-sm font-semibold ${selectedCategoryVariance >= 0 ? 'text-success' : 'text-danger'}`}>
-                    {formatCurrency(selectedCategoryVariance)}
-                  </dd>
-                </div>
-              </dl>
-              {selectedCategoryId !== 'all' && (
-                <button
-                  type="button"
-                  onClick={() => focusCategory(selectedCategoryId, true)}
-                  className="mt-4 text-[11px] font-semibold text-accent hover:text-accent/80"
-                >
-                  Focus in navigator
-                </button>
-              )}
-            </div>
-
-            <div className="rounded-xl border border-slate-800 bg-slate-950/60 p-4">
-              <p className="text-xs font-semibold uppercase tracking-wide text-slate-400">Overspending hotspots</p>
-              {overspendingCategories.length > 0 ? (
-                <ul className="mt-3 space-y-2 text-xs">
-                  {overspendingCategories.map(({ category, summary }) => (
-                    <li key={category.id} className="rounded-lg border border-danger/40 bg-danger/5 px-3 py-2">
-                      <div className="flex items-center justify-between gap-2">
-                        <span className="font-semibold text-slate-100">{category.name}</span>
-                        <span className="font-semibold text-danger">{formatCurrency(summary.variance)}</span>
-                      </div>
-                      <div className="mt-1 flex items-center justify-between gap-2 text-[10px] text-danger/80">
-                        <span>
-                          Spent {formatCurrency(summary.actual)} of {formatCurrency(summary.planned)}
-                        </span>
-                        <button
-                          type="button"
-                          onClick={() => {
-                            focusCategory(category.id, true);
-                            setSelectedCategoryId(category.id);
-                          }}
-                          className="font-semibold text-danger hover:underline"
-                        >
-                          Inspect
-                        </button>
-                      </div>
-                    </li>
-                  ))}
-                </ul>
-              ) : (
-                <p className="mt-3 text-xs text-slate-500">No overspending recorded in this period.</p>
-              )}
-            </div>
-
-            <div className="rounded-xl border border-slate-800 bg-slate-950/60 p-4">
-              <p className="text-xs font-semibold uppercase tracking-wide text-slate-400">Baseline controls</p>
-              {focusedCategory ? (
-                <form onSubmit={handleBudgetSubmit} className="mt-3 space-y-3">
-                  <div>
-                    <p className="text-sm font-semibold text-slate-100">{focusedCategory.name}</p>
-                    <p className="text-[11px] text-slate-500">
-                      {baselineLabel}
-                    </p>
-                  </div>
-                  <input
-                    type="number"
-                    min={0}
-                    step="any"
-                    value={budgetDraft}
-                    onChange={(event) => handleBudgetDraftChange(event.target.value)}
-                    className="w-full rounded-lg border border-slate-700 bg-slate-950/70 px-3 py-2 text-sm text-slate-100 focus:border-accent focus:outline-none"
-                    placeholder="0"
-                  />
-                  <div className="flex flex-wrap gap-2">
-                    <button
-                      type="submit"
-                      className="rounded-lg bg-success px-3 py-2 text-xs font-semibold text-slate-900 transition hover:bg-emerald-400 disabled:opacity-60"
-                      disabled={isSavingBudget}
-                    >
-                      {isSavingBudget ? 'Saving…' : 'Save baseline'}
-                    </button>
-                    <button
-                      type="button"
-                      onClick={() => void handleBudgetClear()}
-                      className="rounded-lg border border-slate-700 px-3 py-2 text-xs font-semibold text-slate-300 transition hover:border-slate-500"
-                      disabled={isSavingBudget}
-                    >
-                      Clear
-                    </button>
-                  </div>
-                  {budgetError && <p className="text-[10px] text-danger">{budgetError}</p>}
-                  {budgetFeedback === 'saved' && <p className="text-[10px] text-success">Baseline saved.</p>}
-                  {budgetFeedback === 'cleared' && <p className="text-[10px] text-slate-400">Baseline cleared for this category.</p>}
-                </form>
-              ) : (
-                <p className="mt-3 text-xs text-slate-500">Select a category in the navigator to set a spending baseline.</p>
-              )}
-            </div>
-          </div>
-
-          <div className="grid gap-4 md:grid-cols-2 xl:grid-cols-4">
-            <div className="space-y-2">
-              <label className="text-xs font-semibold uppercase tracking-wide text-slate-400" htmlFor="budgeting-summary-scope">
-                Summary scope
-              </label>
-              <select
-                id="budgeting-summary-scope"
-                value={selectedCategoryId}
-                onChange={(event) => handleSummaryScopeChange(event.target.value as 'all' | string)}
-                className="w-full rounded-lg border border-slate-700 bg-slate-950/70 px-3 py-2 text-sm text-slate-100 focus:border-accent focus:outline-none"
-              >
-                <option value="all">All categories</option>
-                {categoryOptions.map((option) => (
-                  <option key={option.id} value={option.id}>
-                    {option.label}
-                  </option>
-                ))}
-              </select>
-            </div>
-            <div className="space-y-2">
-              <span className="text-xs font-semibold uppercase tracking-wide text-slate-400">Navigator filter</span>
-              <div className="flex flex-wrap gap-2">
-                {navigatorFilterOptions.map((option) => (
-                  <button
-                    key={option.key}
-                    type="button"
-                    onClick={() => setNavigatorFilter(option.key)}
-                    className={`rounded-full px-3 py-1 text-[11px] font-semibold transition ${
-                      navigatorFilter === option.key
-                        ? 'bg-accent text-slate-900'
-                        : 'bg-slate-800 text-slate-300 hover:bg-slate-700'
-                    }`}
-                  >
-                    {option.label}
-                  </button>
-                ))}
-              </div>
-            </div>
-            <div className="space-y-2">
-              <span className="text-xs font-semibold uppercase tracking-wide text-slate-400">Navigator view</span>
-              <div className="flex flex-wrap gap-2">
-                {navigatorViewOptions.map((option) => (
-                  <button
-                    key={option.key}
-                    type="button"
-                    onClick={() => setNavigatorView(option.key)}
-                    className={`rounded-full px-3 py-1 text-[11px] font-semibold transition ${
-                      navigatorView === option.key
-                        ? 'bg-accent text-slate-900'
-                        : 'bg-slate-800 text-slate-300 hover:bg-slate-700'
-                    }`}
-                  >
-                    {option.label}
-                  </button>
-                ))}
-              </div>
-            </div>
-            <div className="space-y-2">
-              <label className="text-xs font-semibold uppercase tracking-wide text-slate-400" htmlFor="budgeting-search">
-                Search
-              </label>
-              <div className="flex flex-wrap gap-2">
-                <input
-                  id="budgeting-search"
-                  type="search"
-                  placeholder="Search categories or items"
-                  value={categorySearchTerm}
-                  onChange={(event) => setCategorySearchTerm(event.target.value)}
-                  className="min-w-[200px] flex-1 rounded-lg border border-slate-700 bg-slate-950/70 px-3 py-2 text-sm text-slate-100 focus:border-accent focus:outline-none"
-                />
-                <button
-                  type="button"
-                  onClick={handleResetFilters}
-                  className="rounded-lg border border-slate-700 px-3 py-2 text-xs font-semibold text-slate-300 transition hover:border-slate-500"
-                >
-                  Reset
-                </button>
-              </div>
-            </div>
-          </div>
-
-          <div className="flex flex-col gap-6 lg:flex-row lg:items-start lg:justify-between">
-          <div className="space-y-4">
-            {navigatorView === 'category' ? (
-              <>
-                {renderedCategorySections.length > 0 && (
-                  <div className="overflow-hidden rounded-2xl border border-slate-800/70 bg-slate-950/40 shadow-inner">
-                    <div className="overflow-x-auto">
-                      <div className="min-w-[980px] text-xs">
-                        <div className="grid grid-cols-[minmax(0,2.6fr)_minmax(110px,0.9fr)_minmax(120px,0.9fr)_minmax(120px,0.9fr)_minmax(120px,0.9fr)_minmax(220px,1fr)] items-center gap-4 border-b border-slate-800/80 bg-slate-950 px-4 py-3 text-[11px] font-semibold uppercase tracking-wide text-slate-400">
-                          <span className="text-left">Category / Item</span>
-                          <span>Due</span>
-                          <span className="text-right">Planned</span>
-                          <span className="text-right">Spent</span>
-                          <span className="text-right">Variance</span>
-                          <span className="text-right">Workflow</span>
-                        </div>
-                        <div>{renderedCategorySections}</div>
-                      </div>
-                    </div>
-                  </div>
-                )}
-                {visibleUncategorisedDetails.length > 0 && (
-                  <div className="rounded-xl border border-slate-800 bg-slate-950/70 p-4">
-                    <div className="flex flex-col gap-2 sm:flex-row sm:items-center sm:justify-between">
-                      <div>
-                        <h4 className="text-sm font-semibold text-slate-100">Uncategorised items</h4>
-                        <p className="text-[11px] text-slate-500">
-                          Assign a category so these expenses roll into the right budgets.
-                        </p>
-                      </div>
-                      <span className="rounded-full bg-warning/20 px-2 py-0.5 text-[10px] font-semibold text-warning">
-                        {visibleUncategorisedDetails.length} item{visibleUncategorisedDetails.length === 1 ? '' : 's'}
-                      </span>
-                    </div>
-                    <div className="mt-3 space-y-3">
-                      {visibleUncategorisedDetails.map((detail) => renderItemCard(detail, 0))}
-                    </div>
-                  </div>
-                )}
-              </>
-            ) : (
-              <>
-                {PRIORITY_ORDER.map((priority) => {
-                  const items = priorityGroups[priority];
-                  if (items.length === 0) {
-                    return null;
-                  }
-                  const token = PRIORITY_TOKEN_STYLES[priority];
-                  return (
-                    <section
-                      key={priority}
-                      className="rounded-xl border border-slate-800 bg-slate-950/70 p-4"
-                    >
-                      <div className="flex flex-col gap-2 sm:flex-row sm:items-center sm:justify-between">
-                        <div className="flex items-center gap-2">
-                          <span
-                            className={`rounded-full px-2 py-0.5 text-[10px] font-semibold ${token.badgeClass}`}
-                          >
-                            {token.label}
-                          </span>
-                          <span className="text-xs text-slate-500">
-                            {items.length} item{items.length === 1 ? '' : 's'}
-                          </span>
-                        </div>
-                      </div>
-                      <div className="mt-3 space-y-3">
-                        {items.map((detail) => renderItemCard(detail, 0))}
-                      </div>
-                    </section>
-                  );
-                })}
-              </>
-            )}
-            {!hasNavigatorResults && (
-              <p className="text-sm text-slate-500">
-                No planned expenses match your filters yet. Adjust the filters or add new planned spends.
-              </p>
-            )}
-          </div>
-          <aside className="flex flex-col gap-4 rounded-xl border border-slate-800 bg-slate-950/70 p-4 text-sm">
-            {inspectorCategory ? (
-              <>
-                <div>
-                  <h5 className="text-xs font-semibold uppercase tracking-wide text-slate-400">Overspending watchlist</h5>
-                  {inspectorOverspendingItems.length > 0 ? (
-                    <ul className="mt-2 space-y-2">
-                      {inspectorOverspendingItems.map((detail) => (
-                        <li
-                          key={detail.item.id}
-                          className="rounded-lg border border-danger/40 bg-danger/5 px-3 py-2 text-xs text-danger"
-                        >
-                          <div className="flex items-center justify-between gap-2">
-                            <span className="font-semibold">{detail.item.name}</span>
-                            <span>{formatCurrency(Math.abs(detail.variance))}</span>
-                          </div>
-                          <div className="mt-1 flex items-center justify-between gap-2 text-[11px] text-danger/80">
-                            <span>
-                              Spent {formatCurrency(detail.actual)} of {formatCurrency(detail.item.plannedAmount)}
-                            </span>
-                            <button
-                              type="button"
-                              onClick={() => handleFocusDetail(detail)}
-                              className="font-semibold text-danger hover:underline"
-                            >
-                              Adjust
-                            </button>
-                          </div>
-                        </li>
-                      ))}
-                    </ul>
-                  ) : (
-                    <p className="mt-2 text-xs text-slate-500">No overspending yet. Keep tracking!</p>
-                  )}
-                </div>
-                <div>
-                  <h5 className="text-xs font-semibold uppercase tracking-wide text-slate-400">Upcoming items</h5>
-                  {inspectorUpcomingItems.length > 0 ? (
-                    <ul className="mt-2 space-y-2">
-                      {inspectorUpcomingItems.map((detail) => (
-                        <li
-                          key={detail.item.id}
-                          className="rounded-lg border border-slate-800 bg-slate-950/80 px-3 py-2 text-xs text-slate-300"
-                        >
-                          <div className="flex items-center justify-between gap-2">
-                            <span className="font-semibold text-slate-100">{detail.item.name}</span>
-                            <span>
-                              {detail.item.dueDate
-                                ? new Date(detail.item.dueDate).toLocaleDateString('en-IN', {
-                                    month: 'short',
-                                    day: 'numeric'
-                                  })
-                                : 'No due date'}
-                            </span>
-                          </div>
-                          <div className="mt-1 flex items-center justify-between gap-2">
-                            <span>{formatCurrency(detail.item.plannedAmount)}</span>
-                            <button
-                              type="button"
-                              onClick={() => handleFocusDetail(detail)}
-                              className="font-semibold text-accent hover:underline"
-                            >
-                              Edit plan
-                            </button>
-                          </div>
-                        </li>
-                      ))}
-                    </ul>
-                  ) : (
-                    <p className="mt-2 text-xs text-slate-500">No upcoming items in this category.</p>
-                  )}
-                </div>
-                <div>
-                  <h5 className="text-xs font-semibold uppercase tracking-wide text-slate-400">In-progress items</h5>
-                  {inspectorInProgressItems.length > 0 ? (
-                    <ul className="mt-2 space-y-2">
-                      {inspectorInProgressItems.map((detail) => (
-                        <li
-                          key={detail.item.id}
-                          className="rounded-lg border border-slate-800 bg-slate-950/80 px-3 py-2 text-xs text-slate-300"
-                        >
-                          <div className="flex items-center justify-between gap-2">
-                            <span className="font-semibold text-slate-100">{detail.item.name}</span>
-                            {statusBadge(detail.item.status)}
-                          </div>
-                          <div className="mt-1 flex items-center justify-between gap-2 text-[11px] text-slate-400">
-                            <span>
-                              Due{' '}
-                              {new Date(detail.item.dueDate).toLocaleDateString('en-IN', {
-                                month: 'short',
-                                day: 'numeric'
-                              })}
-                            </span>
-                            <button
-                              type="button"
-                              onClick={() => handleFocusDetail(detail)}
-                              className="font-semibold text-accent hover:underline"
-                            >
-                              Update
-                            </button>
-                          </div>
-                        </li>
-                      ))}
-                    </ul>
-                  ) : (
-                    <p className="mt-2 text-xs text-slate-500">No items currently in progress.</p>
-                  )}
-                </div>
-                <div>
-                  <h5 className="text-xs font-semibold uppercase tracking-wide text-slate-400">Completed items</h5>
-                  {inspectorCompletedItems.length > 0 ? (
-                    <ul className="mt-2 space-y-2">
-                      {inspectorCompletedItems.map((detail) => (
-                        <li
-                          key={detail.item.id}
-                          className="rounded-lg border border-slate-800 bg-slate-950/80 px-3 py-2 text-xs text-slate-300"
-                        >
-                          <div className="flex items-center justify-between gap-2">
-                            <span className="font-semibold text-slate-100">{detail.item.name}</span>
-                            <span className="font-semibold text-success">{formatCurrency(detail.actual)}</span>
-                          </div>
-                          <div className="mt-1 flex items-center justify-between gap-2 text-[11px] text-slate-400">
-                            <span>Planned {formatCurrency(detail.item.plannedAmount)}</span>
-                            <button
-                              type="button"
-                              onClick={() => handleFocusDetail(detail)}
-                              className="font-semibold text-accent hover:underline"
-                            >
-                              Review
-                            </button>
-                          </div>
-                        </li>
-                      ))}
-                    </ul>
-                  ) : (
-                    <p className="mt-2 text-xs text-slate-500">No completed items recorded yet.</p>
-                  )}
-                </div>
-              </>
-            ) : (
-              <div className="rounded-lg border border-slate-800 bg-slate-950/80 p-3 text-xs text-slate-500">
-                Choose a category on the left to review overspending, upcoming work, and progress.
-              </div>
-            )}
-          </aside>
-        </div>
+        <CategoryInspector inspector={inspector} utils={utils} />
       </div>
-      </section>
     </div>
   );
 }
